fix lock mailbox deadlock bug
Locks should now no longer get stuck trying to commit a group (while
ensuring one is open) because it is waiting for itself to message where
to send the persist confirmation to. The solution is to subscribe to
persist notification before (instead of after) adding the lock to the
current group.