fix premature unlock commit bug
By splitting the unlock log operation into two phases that runs before
and after modifying the global lock variable (instead of a single phase
that ran before the global update). The first phase registers the intent
to unlock in the log but don't associate it with a group (i.e., the
unlock can't be committed at this point). After the unlock update is
made in global memory, the second phase moves the unlock to the most
recent group (not necesarily open group) enabling the lock to be
committed.

This fixes and issue where an unlock could be associated with a group
and subsequently be committed (once the log lock is released) before the
unlock is made to global memory. This was an issues as committing the
unlock potentially modifies the lock variable with information about
persist status.