fix data-race when reading original data into the log
Before, a whole page was read into the log as an entry was created. This
is a potential data-race for data that hasn't been changed and becomes a
problem if it actually is a data-race and that data is later marked to
have been changed. This can cause a rollback to an earlier version than
intended.

Now, data is copied into the log as changes are written back when the
data initially becomes dirty in the log.